create new package general for helper components
