--- frontend/apps/demo/components/index.tsx
@@ -1,1 +1,0 @@
-export * from './PrePublishModal'
